refactor: split UI pages and router
