+++ minesweeper-ui/js/pages/GamePage.jsx
@@ -1,0 +1,10 @@
+const { useParams } = ReactRouterDOM;
+
+export default function GamePage() {
+  const { id } = useParams();
+  return (
+    <div className="game-page">
+      <p>{id}</p>
+    </div>
+  );
+}
