+++ minesweeper-ui/js/pages/LoginPage.jsx
@@ -1,0 +1,10 @@
+import { LangContext } from '/js/i18n.js';
+
+export default function LoginPage({ onLogin }) {
+  const { t } = React.useContext(LangContext);
+  return (
+    <div className="login-page">
+      <button onClick={onLogin}>{t.login}</button>
+    </div>
+  );
+}
